Retain fractional interwave-portion of GUS samples
This commits fetches samples from GUS memory returned as a
floating point type containing values that span the 16-bit signed
range. This implementation preserves up to 3 significant figures
of the interface-wave portion previously lost due to bit-shifting.

Before:

  - shifted value 147 -> scales to 18,669
  - shifted value 148 -> scales to 18,796

Although these aptitudes technically span the 16-bit range, they
do so in steps of 127 (which itself represents 7 bits of
information), for an effective dynamic-range of 9 bits.

Retaining the inter-wave portion as as fraction allows us to
recover the full dynamic-range offered by the samples in the data:

  - float value 147.000 -> scales to 18669
  - float value 147.010 -> scales to 18670

These amplitudes span the entire 16-bit range in steps of 1, for
an effective dynamic range of 16 bits.