Dampen the GUS stream's amplitude based on a rolling peak
The GUS accumulates all the channels' output into a
single sample, which can exceed the bounds of a 16-bit
value, causing slight distortion or clipping for these
handful of samples.

This commit tracks the peak accumulated sample, and then
(if needed) scales the signal from that point forward.
It does not 'clip' the values, which would change the
nature of the sound. The ensure that the current peak,
and subsequent future peaks only ever hit the top (or
bottom) edge of the 16-bit dynamic range.